Extract callback type aliases into a GC-allowed block.
It turns out, it's completely OK for a `@nogc` to receive a GC-allowed
function as an argument, you just can't call it. Passing it to some
other routine is fine as well.

Moreover, it's not forbidden to pass a `@nogc` function where a
GC-allowed one is expected. So the only thing that needs to be done is
to put the callback type aliases into a block without `@nogc`, so they
are not marked as such.